+++ packages/rest-api/src/middleware/logger.ts
@@ -20,6 +20,14 @@ export const logger = winston.createLogger({
   level: 'info',
   format: winston.format.combine(
     winston.format.timestamp(),
+    winston.format((info) => {
+      if (typeof info.message === 'object') {
+        Object.assign(info, info.message)
+      }
+      info.msg = info.msg || info.message
+      delete info.message
+      return info
+    })(),
     winston.format.json()
   ),
   transports,