+++ packages/synapse-constants/rollup.config.js
@@ -41,7 +41,7 @@ export default [
       }),
       codecovRollupPlugin({
         enableBundleAnalysis: process.env.CODECOV_TOKEN !== undefined,
-        bundleName: 'synapse-constants',
+        bundleName: '@synapsecns/synapse-constants',
         uploadToken: process.env.CODECOV_TOKEN,
         uploadOverrides: {
           sha: process.env.GH_COMMIT_SHA,