+++ packages/synapse-constants/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 1.6.0 (2024-10-05)
+
+
+### Features
+
+* **synapse-constants:** Refactor with rollup for package build and export [SLT-160] ([#3175](https://github.com/synapsecns/sanguine/issues/3175)) ([32cee8e](https://github.com/synapsecns/sanguine/commit/32cee8e3bb88222e9876b0963effb51d72be31a6))
+* **synapse-interface:** add spectral token ([#2558](https://github.com/synapsecns/sanguine/issues/2558)) ([b936449](https://github.com/synapsecns/sanguine/commit/b9364498b32290615d05eb88542051c8c8d2c0ce))
+
+
+
+
+
 ## [1.5.6](https://github.com/synapsecns/sanguine/compare/synapse-constants@1.5.4...synapse-constants@1.5.6) (2024-10-01)
 
 **Note:** Version bump only for package synapse-constants