+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.40.6](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.40.5...@synapsecns/synapse-interface@0.40.6) (2024-10-05)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.40.5](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.40.4...@synapsecns/synapse-interface@0.40.5) (2024-10-03)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface